% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{curate_clusters}
\alias{curate_clusters}
\title{Curate clusters from seqArchR result}
\usage{
curate_clusters(
  sname,
  use_aggl = "ward.D",
  use_dist = "euclid",
  seqArchR_result,
  iter,
  pos_lab = NULL,
  regularize = TRUE,
  topn = 50,
  use_cutk = 2,
  need_change = NULL,
  change_to = NULL,
  final = FALSE,
  dir_path
)
}
\arguments{
\item{sname}{The sample name}

\item{use_aggl}{The agglomeration method to be used for hierarchical
clustering. This is passed on to `seqArchR::collate_seqArchR_result`.
See argument `aggl_method` in \link[seqArchR]{collate_seqArchR_result}.}

\item{use_dist}{The distance method to be used for hierarchical clustering.
This is passed on to `seqArchR::collate_seqArchR_result`. See argument
`dist_method` in \link[seqArchR]{collate_seqArchR_result}.}

\item{seqArchR_result}{The seqArchR result object.}

\item{iter}{Specify which iteration of the seqArchR result should be used
for obtaining clusters.}

\item{pos_lab}{The position labels}

\item{regularize}{Logical. Specify TRUE if the basis vector comparison
is to be regularized. Requires you to set `topn` which is set to 50 as
default. See argument `regularize` in
\link[seqArchR]{collate_seqArchR_result}.}

\item{topn}{Numeric. The top N features (nucleotide-position pairs)
that will be used for distance computation, rest will be ignored.
See argument `topn` in \link[seqArchR]{collate_seqArchR_result}.}

\item{use_cutk}{Value of K (number of clusters) for cutting the hierarchical
clustering tree.}

\item{need_change}{A list of elements (cluster IDs in the input clusters)
that need re-assignment. Elements}

\item{change_to}{A list of elements (cluster IDs in the input clusters)
to be assigned to those that need re-assignment. In case there is a candidate
that needs to be put into a new, independent cluster of itself, use
0 (as numeric). Both `need_change` and `change_to` should be empty lists if
no re-assignment is to be performed.
#' TO-DO What if two elements from independent clusters are to
be put into a totally new cluster, will using 0 work in this case?}

\item{final}{Logical, set to TRUE or FALSE}

\item{dir_path}{The path to the directory where files will be written}
}
\description{
seqArchR result stores the clusters obtained at every iteration.
It is possible that the previously chosen agglomeration and/or distance
method used with hierarchical clustering does not yield reasonable
clusters. This function enables minor curation of the clusters obtained
from the hierarchical clustering step.
}
\details{
This function helps the user work through the curation in at least three
steps.

1. This function uses hierarchical clustering to obtain a clustering result.
The resulting clustering is visualized by a dendrogram, color-coded cluster
assignments, and corresponding sequence logos. Using this visualization,
the user can identify/estimate the (nearly right) number of clusters to cut
the dendrogram. The first call uses K = 1.

2. Re-call the function with the identified value of K. Look at the
visualization to determine if it is good enough, i.e., it requires minor
re-assignments).

3. Identify cases of cluster assignments that you wish to re-assign to
different clusters. These can be noted as a list and supplied in a
subsequent call to the function.

2. In the final call to the function, set `final = TRUE`, supply the re-
assignments as two lists `need_change` and `change_to`.

The return value of the function is the clustering with re-assignments
executed.
}
