% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_combined_panel.R
\name{form_combined_panel}
\alias{form_combined_panel}
\title{Form a combined panel of three plots}
\usage{
form_combined_panel(iqw_tpm_pl, seqlogos_pl, annot_pl)
}
\arguments{
\item{iqw_tpm_pl}{The IQW-TPM plot generated using
\code{\link{iqw_tpm_plots}}}

\item{seqlogos_pl}{The sequence logos oneplot obtained from
\code{\link{per_cluster_seqlogos}} by setting the argument
\code{`one_plot = TRUE`}}

\item{annot_pl}{The annotations oneplot obtained from
\code{\link{per_cluster_annotations}} by setting the argument
\code{`one_plot = TRUE`}}
}
\value{
This function returns a \code{ggplot} plot object.
}
\description{
For a given sample, this function forms a combined panel of
three plots namely, the IQW-TPM boxplots, cluster sequence logos, and
annotations per cluster. All of these individual plots can be generated
using existing seqArchRplus functions
}
\details{
This functionality requires cowplot. The combined
plot panels are arranged horizontally.
}
\author{
Sarvesh Nikumbh
}
