% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{handle_tc_from_cage}
\alias{handle_tc_from_cage}
\title{Handle writing of tag clusters to disk as BED files.}
\usage{
handle_tc_from_cage(
  sname,
  cage_obj,
  fl_size_up = 500,
  fl_size_down = 500,
  bsgenome,
  dir_path = NULL,
  fname_prefix = NULL,
  fname_suffix = NULL,
  write_to_disk = TRUE,
  ret_seqs = TRUE
)
}
\arguments{
\item{sname}{The sample name.}

\item{cage_obj}{The CAGEexp object from CAGEr.}

\item{fl_size_up, fl_size_down}{Numeric. The size of the flanks in the
upstream and downstream directions.}

\item{bsgenome}{The BSgenome file that will be used to obtain sequences of
the organism from.}

\item{dir_path}{The path to the directory where files will be written. By
default, all BED files are written within a subdirectory named "BED", and
all FASTA files are written within a subdirectory named "FASTA", both
created at the `dir_path` location.}

\item{fname_prefix, fname_suffix}{Specify any prefix or suffix string to be
used in the filename. This can be the organism name etc. Specify without
any separator. By default, an underscore is used as a separator in the
filename.}

\item{write_to_disk}{Logical. Specify TRUE to write files to disk.
More specifically, BED files are written to disk only when this is set to
TRUE. For promoter sequences, FASTA files are written to disk if this arg is
set to TRUE, otherwise not. and a DNAStringSet object is returned if
`ret_seqs` is set to TRUE.}

\item{ret_seqs}{Logical. Specify TRUE if promoter sequences are to be
returned as a DNAStringSet object.}
}
\value{
If `ret_seqs` is TRUE, a DNAStringSet object is returned.
Depending on `write_to_disk`, files are written to disk at the specified
location.
}
\description{
Handle writing of tag clusters to disk as BED files. It can also
return corresponding promoter sequences as a DNAStringSet object or write
them to disk as FASTA files.
}
\details{
You can use the fname_prefix and fname_suffix arguments to specify strings
to be used as prefix and suffix for the files. For example, the organism
name can be used as a prefix for the filename. Similarly, for suffix.
}
