% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_iqw.R
\name{iqw_tpm_plots}
\alias{iqw_tpm_plots}
\title{IQW, TPM plots}
\usage{
iqw_tpm_plots(
  sname,
  dir_path,
  info_df,
  clusts,
  iqw = TRUE,
  tpm = TRUE,
  cons = TRUE,
  txt_size = 12,
  use_suffix = NULL,
  use_prefix = "C",
  wTitle = TRUE
)
}
\arguments{
\item{sname}{Sample name}

\item{dir_path}{The path to the directory where files will be written.}

\item{info_df}{A DataFrame object holding information on the tag clusters.
Expected columns (names) are 'chr', 'start', 'end', 'IQW', 'domTPM', and
'strand'.}

\item{clusts}{List of sequence ids in each cluster.}

\item{iqw}{Logical. Specify TRUE if boxplots of interquantile widths (IQW)
of the tag clusters corresponding to the promoters in each cluster
are to be plotted.}

\item{tpm}{Logical. Specify TRUE if boxplots of TPM values for all clusters
are to be plotted.}

\item{cons}{Logical. Specify TRUE if boxplots of conservation scores
(PhastCons scores) for all clusters. If this is TRUE, an additional column
named 'cons' is expected in the `info_df`.}

\item{txt_size}{Specify text size to be used in the plots.}

\item{use_suffix, use_prefix}{Character. Specify any suffix and/or prefix
you wish to add to the filename.}

\item{wTitle}{Logical. If TRUE, the returned plot will contain a default
title, which is the same as the filename. See details.}
}
\value{
The plot(s) as a ggplot2 object. The order of the plots is IQW,
followed by TPM values (when specified), followed by conservation scores
(when specified).
}
\description{
IQW, TPM plots
}
\details{
The plots are written to a file named
"Sample_<sample_name>_IQW_TPM_Cons_plot.pdf" if all of `iqw`, `tpm`,
and `cons` are set to TRUE. This is also set as the plot title if `wTitle`
is set to TRUE.

All plots are arranged by the IQWs (smallest on top, largest at the bottom),
even iff `iqw` is set to FALSE.
}
