% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{order_clusters_iqw}
\alias{order_clusters_iqw}
\title{Order clusters by median or mean interquantile widths}
\usage{
order_clusters_iqw(sname, clusts, info_df, order_by_median = TRUE)
}
\arguments{
\item{sname}{The sample name}

\item{clusts}{List of sequence ids in each cluster.}

\item{info_df}{The data.frame with all tag clusters information. The
following columns are expected in the data.frame:"chr", "start", "end",
"width", "strand", "score", "nr_ctss", "dominant_ctss", "domTPM",
"IQW", "tpm" and two additional columns based on qLow and qUp used.#'}

\item{order_by_median}{Logical. Whether to order to clusters by their
median (when TRUE) or mean (when FALSE) interquantile widths.}
}
\description{
Order clusters by median or mean interquantile widths
}
