% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{seqs_acgt_image}
\alias{seqs_acgt_image}
\title{Visualize all sequences as an image}
\usage{
seqs_acgt_image(
  sname,
  seqs,
  seqs_ord,
  pos_lab,
  xt_freq,
  yt_freq,
  f_height,
  f_width,
  dir_path
)
}
\arguments{
\item{sname}{The sample name}

\item{seqs}{The sequences}

\item{seqs_ord}{The order of sequences}

\item{pos_lab}{The position labels}

\item{xt_freq}{The frequency of xticks}

\item{yt_freq}{The frequency of yticks}

\item{f_height, f_width}{The height and width for the PNG image.}

\item{dir_path}{Specify the /path/to/directory to store results}
}
\description{
Visualize all sequences as an image
}
