% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{write_seqArchR_cluster_track_bed}
\alias{write_seqArchR_cluster_track_bed}
\title{Write seqArchR cluster information in BED files viewable as browser
tracks}
\usage{
write_seqArchR_cluster_track_bed(
  sname,
  clusts = NULL,
  info_df,
  one_zip_all = TRUE,
  org_name = NULL,
  dir_path = NULL,
  include_in_report = TRUE,
  strand_sep = FALSE
)
}
\arguments{
\item{sname}{Sample name}

\item{clusts}{List of sequence ids in each cluster.}

\item{info_df}{The data.frame holding information to be written to the BED
file}

\item{one_zip_all}{Logical. Specify TRUE when the facility to download BED
files for all clusters with one click is desired, otherwise FALSE. This is
relevant only when include_in_report is TRUE}

\item{org_name}{Organism name}

\item{dir_path}{The path where a special folder named `Cluster_BED_tracks`
(default) is created and all BED files are written inside it}

\item{include_in_report}{Logical. Specify TRUE when this function is invoked
to write BED files to disk *and* provide downloadable links in the HTML
report. The corresponding chunk in the Rmarkdown report should set the
parameter `result='asis'`. By setting this to FALSE, BED files are written
to disk, but no downloadable links are provided. Note: This should be TRUE,
when `one_zip_all` argument is set to TRUE (see below). Requires the package
`xfun`.}

\item{strand_sep}{Logical. Specify TRUE if records for each strand are to
be written in separate BED files.}
}
\description{
Writes the seqArchR clusters are BED tracks for viewing in IGV
or any genome browser
}
\details{
Note: For providing downloadable links in the HTML report, the
complete BED files are encoded into base64 strings and embedded with the
HTML report itself. This considerably increases the size of the HTML file,
and can slow down loading of the HTML file in your browser.
}
