% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{write_tc_bed}
\alias{write_tc_bed}
\title{Write Tag Clusters to a BED file}
\usage{
write_tc_bed(gr, bed_fname)
}
\arguments{
\item{gr}{The GenomicRanges object}

\item{bed_fname}{The BED filename that will be written to disk (with
full path)}
}
\description{
Write Tag Clusters to a BED file
}
\details{
Writes the provided `gr` to the file named `bed_fname`
}
