% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_generate_html_report.R
\name{generate_html_report}
\alias{generate_html_report}
\title{Generate HTML report with scrollable combined panel plots}
\usage{
generate_html_report(
  snames,
  file_type = "PDF",
  img_ht = "1200px",
  img_wd = "1600px",
  page_wd = "1800px",
  render_silently = TRUE,
  dir_path
)
}
\arguments{
\item{snames}{Sample names to be included}

\item{file_type}{"PDF" (default) or "SVG" (to be supported in the future).
The type of files to look for in the sample-specific results folder}

\item{img_ht, img_wd}{The height and width (in pixels) of the images in the
output HTML report. Default values are '1200px' (height) and
'1600px' (width)}

\item{page_wd}{The width of the body in the HTML. Default is '1800px'.}

\item{render_silently}{Logical. TRUE or FALSE}

\item{dir_path}{Specify the `/path/to/directory` where sample-specific
results folders are located. This is a required argument and cannot be NULL.
A directory named `combined_results` is created at the given location, and
the HTML report is written into it}
}
\value{
Nothing. Report is written to disk at the provided \code{`dir_path`
} using the filename \code{
'Combined_panels_report_samples_<samples_names>.html'}.
}
\description{
This function generates an HTML report with large scrollable
combined panels for multiple samples that eases comparison of changes
between samples
}
\details{
This functionality requires suggested libraries \code{
`slickR`} and \code{`pdftools`} installed.
The function assumes requires that the candidate figure files have combined_panel
Note that the combined plot panels are arranged horizontally and
therefore are best viewed in wide desktop monitors.
}
\author{
Sarvesh Nikumbh
}
