% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{per_cluster_annotations}
\alias{per_cluster_annotations}
\title{per_cluster_annotations}
\usage{
per_cluster_annotations(
  sname,
  clusts,
  tc_gr,
  cager_obj = NULL,
  qLow = 0.1,
  qUp = 0.9,
  txdb_obj = NULL,
  tss_region = NULL,
  orgdb_obj = NULL,
  one_plot = TRUE,
  dir_path = NULL,
  txt_size = 12,
  use_suffix = NULL,
  use_prefix = "C",
  n_cores = 1
)
}
\arguments{
\item{sname}{Sample name}

\item{clusts}{List of sequence ids in each cluster.}

\item{tc_gr}{Tag clusters as \code{\link[GenomicRanges]{GRanges}}. If
`cager_obj` is not provided (is NULL), this argument is required.
It will be ignored only if cager_obj is provided.}

\item{cager_obj}{A CAGEexp object obtained from the CAGEr package, if and
when CAGEr was used to process the raw CAGE data}

\item{qLow, qUp}{The interquantile boundaries to be considered for obtaining
tag clusters from the CAGEexp object. See \code{\link[CAGEr]{tagClusters}}}

\item{txdb_obj}{A TxDb object storing transcript metadata}

\item{tss_region}{For ChIPseeker}

\item{orgdb_obj}{Organism-level annotation package}

\item{one_plot}{Logical. Default is TRUE. If set to FALSE the barplots of
annotations per cluster are returned as a list, else all are condensed into
single plot}

\item{dir_path}{Specify the /path/to/directory to store results}

\item{txt_size}{Adjust text size for the plots}

\item{use_suffix, use_prefix}{Character. Specify any suffix and/or prefix
you wish to add to the filename}

\item{n_cores}{Numeric. If you wish to parallelize annotation of peaks,
specify the number of cores. Default is 1 (serial)}
}
\description{
per_cluster_annotations
}
