% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_strand_distributions.R
\name{per_cluster_strand_dist}
\alias{per_cluster_strand_dist}
\title{Visualize as barplots how promoters in each cluster are distributed
on different chromosomes and strands}
\usage{
per_cluster_strand_dist(sname, clusts, info_df, dir_path, colrs = "Paired")
}
\arguments{
\item{sname}{The sample name}

\item{clusts}{List of sequence ids in each cluster.}

\item{info_df}{The information data.frame}

\item{dir_path}{Specify the path to the directory on disk where plots
will be saved}

\item{colrs}{Specify colors used for two strands}
}
\value{
A list of plots showing the per cluster division of promoters on
chromosomes and strands. Plots are also written to disk.
}
\description{
Visualize as barplots how promoters in each cluster are distributed
on different chromosomes and strands
}
