% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR-analysis-functions.R
\name{plot_motif_heatmaps}
\alias{plot_motif_heatmaps}
\title{Plot heatmaps of motifs occuring in seqArchR clusters}
\usage{
plot_motif_heatmaps(
  sname,
  seqs,
  flanks = c(50),
  clusts,
  use_colors = NULL,
  motifs,
  dir_path,
  fheight = 500,
  fwidth = 500,
  funits = "px"
)
}
\arguments{
\item{sname}{The sample name}

\item{seqs}{The sequences as a DNAStringSet object}

\item{flanks}{Flank size. The same flank is used upstream and downstream.
A vector of values is also accepted when more than oone flanks should be
visualized.}

\item{clusts}{List of sequence Ids in each cluster.}

\item{use_colors}{Specify colors to use}

\item{motifs}{A vector of motifs to be visualized in the sequence. This can
be any words formed by the
\href{https://www.bioinformatics.org/sms2/iupac.html}{IUPAC code}.
For example, TATAA, CG, WW, SS etc.}

\item{dir_path}{The path to the directory}

\item{fheight, fwidth, funits}{Height and width of the PDF file, and the units
in which they are specified.}
}
\value{
Nothing. PNG images are written to disk using the provided filenames.
}
\description{
Plot heatmaps of motifs occuring in seqArchR clusters
}
