% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_motif_heatmaps.R
\name{plot_motif_heatmaps}
\alias{plot_motif_heatmaps}
\title{Plot heatmaps of motifs occuring in seqArchR clusters}
\usage{
plot_motif_heatmaps(
  sname,
  seqs,
  flanks = c(50),
  clusts,
  use_colors = NULL,
  motifs,
  dir_path,
  fheight = 500,
  fwidth = 500,
  funits = "px",
  n_cores = 1
)
}
\arguments{
\item{sname}{The sample name}

\item{seqs}{The sequences as a DNAStringSet object}

\item{flanks}{Flank size. The same flank is used upstream and downstream.
A vector of values is also accepted when more than oone flanks should be
visualized.}

\item{clusts}{List of sequence Ids in each cluster.}

\item{use_colors}{Specify colors to use}

\item{motifs}{A vector of motifs to be visualized in the sequence. This can
be any words formed by the
\href{https://www.bioinformatics.org/sms2/iupac.html}{IUPAC code}.
For example, TATAA, CG, WW, SS etc.}

\item{dir_path}{The path to the directory}

\item{fheight, fwidth, funits}{Height and width of the PDF file, and the units
in which they are specified}

\item{n_cores}{Numeric. If you wish to parallelize annotation of peaks,
specify the number of cores. Default is 1 (serial)}
}
\value{
Nothing. PNG images are written to disk using the provided filenames.
}
\description{
Plot heatmaps of motifs occuring in seqArchR clusters
}
\examples{

library(Biostrings)
raw_seqs <- Biostrings::readDNAStringSet(
                          filepath = system.file("extdata",
                            "promoters_small_200.fa.gz",
                            package = "seqArchRplus",
                            mustWork = TRUE)
                        )

use_clusts <- readRDS(system.file("extdata", "clust_info.rds",
         package = "seqArchRplus", mustWork = TRUE))

plot_motif_heatmaps(sname = "sample1", seqs = raw_seqs,
                    flanks = c(50, 100, 200),
                    clusts = use_clusts,
                    motifs = c("WW", "SS", "TATAA", "CG", "Y"),
                    dir_path = tempdir(),
                    fheight = 800, fwidth = 1600)

}
