% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchRplus_acgt_image.R
\name{seqs_acgt_image}
\alias{seqs_acgt_image}
\title{Visualize all sequences as an image}
\usage{
seqs_acgt_image(
  sname,
  seqs,
  seqs_ord,
  pos_lab,
  xt_freq = 5,
  yt_freq = 500,
  f_height = 1200,
  f_width = 600,
  dir_path
)
}
\arguments{
\item{sname}{The sample name}

\item{seqs}{The sequences}

\item{seqs_ord}{The order of sequences}

\item{pos_lab}{The position labels}

\item{xt_freq}{The frequency of xticks}

\item{yt_freq}{The frequency of yticks}

\item{f_height, f_width}{The height and width for the PNG image.}

\item{dir_path}{Specify the /path/to/directory to store results}
}
\value{
Nothing. PNG images are written to disk using the provided filename.
}
\description{
Visualize all sequences as an image
}
\examples{

library(Biostrings)
raw_seqs <- Biostrings::readDNAStringSet(
                          filepath = system.file("extdata",
                            "example_promoters45.fa.gz",
                            package = "seqArchRplus",
                            mustWork = TRUE)
                        )

use_clusts <- readRDS(system.file("extdata", "example_clust_info.rds",
         package = "seqArchRplus", mustWork = TRUE))

seqs_acgt_image(sname = "sample1",
                seqs = raw_seqs,
                seqs_ord = unlist(use_clusts),
                pos_lab = -45:45,
                dir_path = tempdir())

}
\author{
Sarvesh Nikumbh
}
